+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### Unreleased
 
+* Add helper method to get a DateTimeImmutable from a unix timestamp, in current timezone
+
 ### v1.5.0-beta1 (2020-05-14)
 
 * Add AssociativeArrayUtils for common operations on associative arrays